add loading state to signup
