+++ client/src/main.tsx
@@ -1,11 +1,11 @@
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter, Routes, Route} from 'react-router'
 import './index.css'
-import App from './components/App.tsx'
-import LoginSelector from "./components/auth/LoginSelector.tsx";
-import Placeholder from "./components/Placeholder.tsx";
-import LoginPage from "./components/auth/LoginPage.tsx";
-import SignupPage from "./components/auth/SignupPage.tsx";
+import App from './pages/App.tsx'
+import LoginSelector from "./pages/auth/LoginSelector.tsx";
+import Placeholder from "./pages/Placeholder.tsx";
+import LoginPage from "./pages/auth/LoginPage.tsx";
+import SignupPage from "./pages/auth/SignupPage.tsx";
 
 createRoot(document.getElementById('root')!).render(
   <BrowserRouter>
